Update README.md
Some ideas about image compression.

During image compression, we know the ground truth during encoding and decoding. Therefore, how about overfitting each sample during the input. It is like attention, the network is very related with the input. Unlike most cases, the network is isolated and will not adjust the neural's weights according to the input.